ubuntu: fix priority of the artik-platform repository
This patch fixes the priority of the artik-platform repository.
Originally, we intended the artik-platform repository could have the
highest priority than the ubuntu official repository. However, it's not
working because it has full URL address. So, we should change the pin
file can have correct url address of the artik repository.

Change-Id: Ic054998fe4dadd32a35342cda9a8125b79f09417
Signed-off-by: Chanho Park <chanho61.park@samsung.com>